# micro-optimize unique_list