@zzzeek 2015-04-24
I had some recollection that we used a set for this, but apparently it was super long ago and it was this thing:

    def unique_list(seq, compare_with=set):
        seen = compare_with()
        return [x for x in seq if x not in seen and not seen.add(x)]    

hmm yup and that was it until b79a5e76 when it became the current thing.    I'll try running yours in a few.