@zzzeek 2015-04-24
OK so I think also the bench should compare against one with `seen  = {};  seen_add = seen.__setitem__`.   need lunch first