@zzzeek 2015-04-24
yeah.   i ran this all the way back to 2.4 to see why i did it that way, even in 2.4, set is faster.     shrugs!

    def unique_list_setitem(seq):
        seen = {}
        return [
            x for x in seq
            if x not in seen
            and not seen.__setitem__(x, True)]


    def unique_list_add_inline(seq):
        seen = set()
        seen_add = seen.add
        return [
            x for x in seq
            if x not in seen
            and not seen_add(x)]


    def unique_list_setitem_inline(seq):
        seen = {}
        seen_add = seen.__setitem__
        return [
            x for x in seq
            if x not in seen
            and not seen_add(x, True)]


    def unique_list_add(seq):
        seen = set()
        return [x for x in seq if x not in seen and not seen.add(x)]


    import time


    def timeit(fn, number=10000):
        now = time.time()
        for i in xrange(number):
            fn()
        return time.time() - now

    biglist = list(range(10000))
    hugelist = list(range(int(1e6)))
    smalllist = list(range(3))

    for list_ in smalllist, biglist, hugelist:
        batch = []
        for fn in [
            unique_list_add,
            unique_list_setitem,
            unique_list_add_inline,
            unique_list_setitem_inline
        ]:

            num = int(1e6 / len(list_))
            batch.append(
                (fn, timeit(lambda: fn(list_), number=num))
            )

        print("\n")
        for fn, time_ in sorted(batch, key=lambda x: x[1]):
            print(
                "%s, %d, %d times: %s" % (
                    fn.__name__,
                    len(list_),
                    num,
                    time_
                )

            )