@zzzeek 2015-04-24
here's 2.7:

	unique_list_add_inline, 3, 333333 times: 0.347361803055
	unique_list_add, 3, 333333 times: 0.348173141479
	unique_list_setitem_inline, 3, 333333 times: 0.410316944122
	unique_list_setitem, 3, 333333 times: 0.426838874817


	unique_list_add_inline, 10000, 100 times: 0.159450054169
	unique_list_add, 10000, 100 times: 0.189424037933
	unique_list_setitem_inline, 10000, 100 times: 0.227113962173
	unique_list_setitem, 10000, 100 times: 0.268784999847


	unique_list_add_inline, 1000000, 1 times: 0.204242229462
	unique_list_add, 1000000, 1 times: 0.241760969162
	unique_list_setitem_inline, 1000000, 1 times: 0.288633108139
	unique_list_setitem, 1000000, 1 times: 0.342014074326