@gdementen 2015-04-24
You're welcome, I am using that code in an unrelated project so it's only fair I contribute back improvements. FWIW, I had timed it on both 3.4 and 2.7.